Continue reco functions re-arrangement
