+++ fanal/reco/reco_functions.py
@@ -125,175 +125,3 @@ def reconstruct_event(det_name          : DetName,
                          f"E: {voxel.E/units.keV:5.1f} keV")
 
     return event_data
-
-
-
-############################  POSITION FUNCTIONS  ############################
-
-def translate_hit_positions(detname        : DetName,
-                            active_mcHits  : pd.DataFrame,
-                            drift_velocity : float
-                           ) -> None:
-    """
-    In MC simulations all the hits of a MC event are assigned to the same event.
-    In some special cases these events may contain hits in a period of time
-    much longer than the corresponding to an event.
-    Some of them occur with a delay that make them being reconstructed in shifted-Zs.
-    This functions accomplish all these situations.
-    """
-
-    # Minimum time offset to fix
-    min_offset = 1.*units.mus
-
-    hit_times = active_mcHits.time
-    min_time, max_time = hit_times.min(), hit_times.max()
-    
-    # Only applying to events wider than a minimum offset
-    if ((max_time - min_time) > min_offset):
-        if is_detector_symmetric(detname):
-            active_mcHits['shifted_z'] = active_mcHits['z'] - np.sign(active_mcHits['z']) * \
-                                         (active_mcHits['time'] - min_time) * drift_velocity
-        # In case of assymetric detector
-        else:
-            active_mcHits['shifted_z'] = active_mcHits['z'] + \
-                                         (active_mcHits['time'] - min_time) * drift_velocity
-
-    # Event time width smaller than minimum offset        
-    else:
-        active_mcHits['shifted_z'] = active_mcHits['z']
-
-
-
-def check_event_fiduciality(detname      : DetName,
-                            veto_width   : float,
-                            min_VetoE    : float,
-                            event_voxels : List[Voxel]
-                           ) -> Tuple[float, float, float, float, bool]:
-    """
-    Checks if an event is fiducial or not.
-    
-    Parameters:
-    -----------
-    detname      : DetName
-        Name of the detector.
-    veto_width   : float
-        Width of the veto region.
-    min_VetoE    : float
-        Veto energy threshold.
-    event_voxels : List[Voxel]
-        List of voxels of the event.
-        
-    Returns:
-    --------
-    Tuple of 5 values containing:
-        3 floats containing voxels minZ, maxZ and maxRad
-        A float with the nergy deposited in veto
-        A bool saying if the event is fiducial or not.
-    """
-    
-    voxels_Z   = [voxel.Z for voxel in event_voxels]
-    voxels_Rad = [(math.sqrt(voxel.X**2 + voxel.Y**2)) for voxel in event_voxels]
-
-    voxels_minZ   = min(voxels_Z)
-    voxels_maxZ   = max(voxels_Z)
-    voxels_maxRad = max(voxels_Rad)
-
-    fid_dimensions = get_fiducial_size(detname, veto_width)
-    
-    if is_detector_symmetric(detname):
-        veto_energy = sum(voxel.E for voxel in event_voxels \
-                          if ((voxel.Z < fid_dimensions.z_min) |
-                              (voxel.Z > fid_dimensions.z_max) |
-                              (math.sqrt(voxel.X**2 + voxel.Y**2) > fid_dimensions.rad) |
-                              # Extra veto for central cathode
-                              (abs(voxel.Z) < veto_width)))
-    else:
-        veto_energy = sum(voxel.E for voxel in event_voxels \
-                          if ((voxel.Z < fid_dimensions.z_min) |
-                              (voxel.Z > fid_dimensions.z_max) |
-                              (math.sqrt(voxel.X**2 + voxel.Y**2) > fid_dimensions.rad)))
-        
-    fiducial_filter = veto_energy < min_VetoE
-
-    return voxels_minZ, voxels_maxZ, voxels_maxRad, veto_energy, fiducial_filter
-
-
-
-############################   ENERGY FUNCTIONS   ############################
-
-# Some constants needed
-S1_Eth    = 20. * units.keV  # Energy threshold for S1s
-S1_WIDTH  = 10. * units.ns   # S1 time width
-EVT_WIDTH =  5. * units.ms   # Recorded time per event
-
-# Recorded time per event in S1_width units
-evt_width_inS1 = EVT_WIDTH / S1_WIDTH
-
-
-
-def get_S1_E(evt_hits: pd.DataFrame
-	        ) -> List[Tuple[float,float]]:
-	"""
-	It returns a List of tuples containing:
-	(S1 time, S1 energy) for all the S1s
-	inside the recorded_time per event
-	"""
-	# Discretizing hit times in s1_widths
-	evt_t0 = evt_hits.time.min()
-	evt_hits['disc_time'] = ((evt_hits['time'] - evt_t0) // S1_WIDTH) * S1_WIDTH
-
-	# Getting S1s Info
-	S1s = []
-
-	s1_ts = evt_hits[evt_hits.disc_time <= evt_width_inS1]['disc_time'].unique()
-	for s1_t in s1_ts:
-		s1_E = evt_hits[evt_hits.disc_time==s1_t].energy.sum()
-		S1s.append((s1_t, s1_E))
-
-	return S1s
-
-
-
-def get_mc_energy(evt_hits: pd.DataFrame
-				 ) -> float:
-	"""
-	It returns the mc energy of an event if there is only one S1.
-	It returns 0 in any other case.
-	"""
-	S1_Es   = get_S1_E(evt_hits)
-	logger.debug(f"  Num S1s:  {len(S1_Es)}")
-	logger.debug(f"  S1 collection (S1_t, S1_E):  {S1_Es}")
-
-	# Checking the number of S1s with E > Eth,
-	# and collecting the energy
-	num_S1s = 0
-	evt_mcE = 0.
-	for i in range(len(S1_Es)):
-		S1_E = S1_Es[i][1]
-		if S1_E > S1_Eth:
-			num_S1s += 1
-			if (num_S1s == 1): 
-				evt_mcE = S1_E
-			# If more than 1 S1 with E > Eth -> Set evt_mcE to 0
-			else:
-				evt_mcE = 0.
-				break
-		# Collecting the energy of S1 with E < Eth
-		else:
-			evt_mcE += S1_E
-        
-	return evt_mcE
-
-
-
-def smear_evt_energy(mcE       : float,
-                     sigma_Qbb : float,
-                     Qbb       : float
-					) -> float:
-	"""
-	It smears and returns the montecarlo energy of the event (mcE) according to:
-	the sigma at Qbb (sigma_Qbb) in absolute values (keV).
-	"""
-	sigma_E = sigma_Qbb * math.sqrt(mcE/Qbb)
-	smE = np.random.normal(mcE, sigma_E)
-	return smE
